Add files via upload
Review Analyzer to classify positive and negative reviews of the customers. This dataset is based on amazon customers who have written reviews on food products. This dataset was available with an online course offered on edx.org by MITx.

This analyzer is purely a case study based upon various algorithms learned in online course. 
Made a strategy to convert plan text to feature vectors and then train model.

1. load dataset by using load_data() function.
2. treat each review as one item.
3. extract words from available data and create a dictionary without punctuations and stop words. 
4. Create feature vectors for each review based on dictionary.